Add new way to switch classes
The url of the current page is stored in the url const.
A target url is stored in the path const.
The 2 consts are compared against each other in a ternary condition to
know which CSS rule to set, instead of the use of a state.
The "exact" attribute allows to perfectly match the path set in the "to"
attribute.
The "activeClassName" is activated only when the page is rendered.
Removed the "onClick" attribute as it is no longer needed.