[refactor] Set new API call to better handle tags
Once the photos have been fetched and stored in an array, a same amount of promises is created and launched to get all tags linked to the image via the ID.

With the requests over, the useful data from the two arrays is reunited in a single array and set in a state.
In the same time, the final result is stored in a session storage in order to avoid new calls to the API if the session is still on.